requirements.txt
requests
openai
google-api-python-client
google-auth
google-auth-oauthlib
beautifulsoup4